backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
google-genai>=1.10.0
parallel-web>=0.1.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.10.0
